% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/do.cut.R
\name{do.cut}
\alias{do.cut}
\title{Apply Cuts to the Reference Matrix}
\usage{
do.cut(mat, cuts, type = "combined")
}
\arguments{
\item{mat}{the matrix to cut}

\item{cuts}{a list of cuts generated using \code{\link{make.cut}}}

\item{type}{the type of cuts to use (use \code{combined} by default)}
}
\description{
Apply cuts generated using the \code{\link{make.cut}} function to the reference matrix
}
\details{
The matrix can be cut using either the fixed cuts (\code{type='fixed'}), or the combined cuts (\code{type='combined'})
where the limits have been adjusted to match local minima and maxima.
}
\examples{
# generate a random 3D matrix
mat <- matrix(rnorm(300),ncol=3)
dimnames(mat)[[2]] <- LETTERS[1:3]
# generate 2 bins with a minimum bin size of 5
cuts <- make.cut(mat,n=3,count.lim=5)
par(mfrow=n2mfrow(ncol(mat)),
mar=c(1,1,3,1))
ksink <- lapply(dimnames(mat)[[2]],function(cur.ch) {
   plot(cuts[[cur.ch]]$dens,
        main=paste('Fixed Cuts for',cur.ch))
   abline(v=cuts[[cur.ch]]$fixed,lty=2,col=2)
 }
)
# Generate the cuts
cut.mat <- do.cut(mat,cuts,type='fixed')
head(cut.mat)
}
\author{
Yann Abraham
}

