% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/js.dist.R
\name{js.dist}
\alias{js.dist}
\title{Compute the Jensen-Shannon Distance between 2 sets of Hilbert Index}
\usage{
js.dist(mat, pc = 1e-04)
}
\arguments{
\item{mat}{a matrix of counts, where rows correspond to sub-cubes and columns to samples}

\item{pc}{a pseudo-count that is added to all samples to avoid divide-by-zero errors}
}
\value{
a S3 distance object
}
\description{
The \href{https://en.wikipedia.org/wiki/Jensen-Shannon_divergence}{Jensen-Shannon distance} is a method to
measure the distance between discrete probability distributions. To measure the distance between 2 high-dimensional
datasets, we cut the space into sub-cubes, then count the number of events per cube. The resulting probability
distributions can be compared using the Jensen-Shannon distance.
}
\examples{
# generate 3 samples over 5 dimensions
# sample 1 and 2 are similar, sample 3 has an extra population
my.samples <- lapply(LETTERS[1:3],function(x) {
    # each sample has a different number of events
    cur.mat <- matrix(rnorm(5*floor(runif(1)*10000)),ncol=5)
    dimnames(cur.mat)[[2]] <- LETTERS[(length(LETTERS)-4):length(LETTERS)]
    return(cur.mat)
  }
)
names(my.samples) <- LETTERS[1:3]
# add an extra population to sample C
my.samples[['C']] <- rbind(my.samples[['C']],
                           matrix(rnorm(5*floor(runif(1)*100),2),ncol=5))
# assemble a sample matrix
my.samples.mat <- do.call('rbind',my.samples)
my.samples.id <- lapply(names(my.samples),
                        function(cur.spl) rep(cur.spl,nrow(my.samples[[cur.spl]])))
my.samples.id <- unlist(my.samples.id)

# cut it
my.cuts <- make.cut(my.samples.mat,n=5,count.lim=5)
my.samples.cut <- do.cut(my.samples.mat,my.cuts,type='fixed')

# compute the hilbert index
system.time(my.samples.index <- do.hilbert(my.samples.mat,horder=4))

# assemble a contingency table
my.samples.table <- table(my.samples.index,my.samples.id)
dim(my.samples.table)

heatmap(log10(my.samples.table+0.00001),
        col=colorRampPalette(c('white',blues9))(24),
        Rowv=NA,Colv=NA,
        scale='none')

# compute the Jensen-Shannon distance
my.samples.dist <- js.dist(my.samples.table)
my.samples.clust <- hclust(my.samples.dist)

plot(my.samples.clust)
}
\author{
Yann Abraham
}

