% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localMinima.R
\name{localMinima}
\alias{localMinima}
\title{Find Local Minima in a vector}
\usage{
localMinima(x)
}
\arguments{
\item{x}{a vector of density values, as generated through a call to \code{\link{density}}}
}
\value{
a vector of index corresponding to local minima
}
\description{
Given a density object, find the position of local minima (inflection points)
}
\examples{
x <- c(rnorm(100),rnorm(100,3))
dx <- density(x)
plot(dx)
abline(v=dx$x[localMinima(dx$y)],col=2,lty=2)
}
\author{
Tommy \url{http://stackoverflow.com/questions/6836409/finding-local-maxima-and-minima}
}
